behavior subject

+++ src/app/components/admin/admin.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
+import { Router } from '@angular/router';
 import { User } from 'src/app/entities/user.model';
 import { AuthenticationService } from 'src/app/services/authentication.service';
 import { UserService } from 'src/app/services/user.service';
@@ -16,30 +17,39 @@ export class AdminComponent implements OnInit {
   errorMessage: string = "";
   displayedColumns: string[] = ['id', 'username', 'name', 'role', 'create_time', 'last_login', 'actions']
 
-  @ViewChild(UserComponent) child: UserComponent | undefined;
-  constructor(private userService: UserService) { }
+  constructor(private userService: UserService, private router: Router) { }
 
   ngOnInit(): void {
     this.userService.getAllUsers().subscribe(
       data => {
         this.userList = data;
       }
-    )
+    );
   }
 
   editUser(user: User){
-    
+    this.userService.editedUserSubject.next(user);
   }
 
   deleteUser(user: User, index: number) {
-    this.userService.deleteUser(user).subscribe(
-      data => {
-        this.userList.splice(index, 1);
-      }, err => {
-        this.errorMessage = "Neočekávaná chyba";
-        console.log(err);
-      }
-    )
+    if(confirm("Potvrzeni: bude smazán uživatel " + user.username)){
+      this.userService.deleteUser(user).subscribe(
+        data => {
+          this.userList.splice(index, 1);
+        }, err => {
+          this.errorMessage = "Neočekávaná chyba";
+          console.log(err);
+        }
+      );
+      this.reloadComponent();
+    }
   }
+
+  reloadComponent() {
+    let currentUrl = this.router.url;
+        this.router.routeReuseStrategy.shouldReuseRoute = () => false;
+        this.router.onSameUrlNavigation = 'reload';
+        this.router.navigate([currentUrl]);
+    }
 
 }
